refactor(agent): step 抽象做满——步骤独立成文件、规划也成 StepHandler
把「可插拔 step」落到实处（行为不变，49 tests 全绿）：
- StepHandler<Ctx, R = void> 泛化出返回类型，规划单轮得以成为正式步骤。
- steps/review.ts：describe-review / judge / asks / summary 四个 StepHandler + REVIEW_STEPS
  注册表 + 判读/总结提示词与解析，全部从 orchestrator 移出独立维护。
- steps/planning.ts：plan-cycle 抽为 planCycleStep（StepHandler<…, PlanCycleOutcome>）+ 协议/
  会话上下文/记忆/计划等助手，从 planner 移出。
- orchestrator.ts / planner.ts 退化为「装配 + 跑步骤」的瘦驱动；json/labels/类型仍留原处（测试
  与外部按原路径引用），与 step 文件经运行期安全的引用环协作。
